Apply patches to non working essentials features
